Refactor: Move deposit tx related methods to DepositTxValidation
