+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/buyer/BuyerVerifiesPreparedDelayedPayoutTx.java
@@ -41,7 +41,7 @@ import lombok.extern.slf4j.Slf4j;
 
 import static bisq.core.trade.validation.DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount;
 import static bisq.core.trade.validation.DelayedPayoutTxValidation.checkLockTime;
-import static bisq.core.trade.validation.TradeValidation.checkRawTransactionInputsAreNotMalleable;
+import static bisq.core.trade.validation.DepositTxValidation.checkRawTransactionInputsAreNotMalleable;
 import static bisq.core.trade.validation.TradeValidation.toVerifiedTransaction;
 import static com.google.common.base.Preconditions.checkNotNull;
