+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/buyer_as_taker/BuyerAsTakerSendsDepositTxMessage.java
@@ -33,7 +33,7 @@ import java.util.UUID;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.validation.TradeValidation.checkTransactionIsUnsigned;
+import static bisq.core.trade.validation.DepositTxValidation.checkTransactionIsUnsigned;
 
 
 @Slf4j
