+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/seller_as_maker/SellerAsMakerProcessDepositTxMessage.java
@@ -26,7 +26,7 @@ import bisq.common.taskrunner.TaskRunner;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.validation.TradeValidation.checkTransactionIsUnsigned;
+import static bisq.core.trade.validation.DepositTxValidation.checkTransactionIsUnsigned;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
